backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
yfinance
pandas
beautifulsoup4
feedparser
openai
# For potential local AI usage
# langchain
google-generativeai
pandas_ta
colorama
